requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.6
Django==4.1.7
django-environ==0.9.0
django-leaflet==0.28.3
django-macros==0.4.0
django-modeladmin-reorder==0.3.1
django-postgres-copy==2.7.2
django_debug_toolbar==3.8.1
idna==3.4
joblib==1.2.0
nltk==3.8.1
numpy==1.24.2
pandas==1.5.3
pip-autoremove==0.10.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
regex==2023.6.3
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
sqlparse==0.4.3
threadpoolctl==3.1.0
tomotopy==0.12.4
tqdm==4.65.0
urllib3==1.26.14
